RLIB-27: No GLIB types in rlib.i
Regenerated language bindings.

+++ bindings/interfaces/rlibJNI.java
@@ -44,7 +44,7 @@ public class rlibJNI {
   public final static native int rlib_graph_clear_bg_region(long jarg1, String jarg2);
   public final static native int rlib_graph_set_x_minor_tick(long jarg1, String jarg2, String jarg3);
   public final static native int rlib_graph_set_x_minor_tick_by_location(long jarg1, String jarg2, int jarg3);
-  public final static native long rlib_graph(long jarg1, long jarg2, long jarg3, double jarg4, long jarg5);
+  public final static native double rlib_graph(long jarg1, long jarg2, long jarg3, double jarg4, long jarg5);
   public final static native int rlib_parse(long jarg1);
   public final static native void rlib_set_query_cache_size(long jarg1, int jarg2);
   public final static native void rlib_set_numeric_precision_bits(long jarg1, int jarg2);
